requirements.txt
pandas>=2.1.0
requests>=2.31.0
gspread>=5.7.2
google-auth>=2.23.0
python-dotenv>=1.0.0
